Add interactive project initialization with agent and RAG setup
Includes new src/init.rs for guided config, agent presets, RAG options,
and demo document creation. Integrates init module into main.rs.
