+++ .squidignore.template
